refactor : 프로젝트 폴더 구조 리팩토링
