--- lib/member/screens/login_screen.dart
@@ -1,17 +1,0 @@
-import 'package:flutter/material.dart';
-
-class LoginScreen extends StatelessWidget {
-  const LoginScreen({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text("Login"),
-      ),
-      body: const Center(
-        child: Text("Login 페이지입니다."),
-      ),
-    );
-  }
-}
