+++ lib/sentence/repository/sentence_repository.dart
@@ -1,6 +1,6 @@
 // 데이터 처리 로직
 
-import 'package:typed/sentence_collection/data/models/sentence_model.dart';
+import 'package:typed/sentence/model/sentence_model.dart';
 
 class SentenceRepository {
   final List<Sentence> _sentences = []; // 임시 저장소 (나중에 DB로 대체 가능)
